Renamed things to be per user and not hidden
Renamed to .lib/ to lib/

Renamed aoc.ys to personal ingydotnet.ys

Added a lib/README.md

Note: Every YS solution I contribute that needs AOC data will use that
file. So I thought it best to put it in a general place and not repeat
it in every solution dir.